requirements.txt
anthropic
python-dotenv
flask
chromadb
